+++ scoreboard/src/test/java/net/apartium/cocoabeans/scoreboard/display/packet/RemoveTeamPacket.java
@@ -1,6 +1,6 @@
 package net.apartium.cocoabeans.scoreboard.display.packet;
 
-import net.apartium.cocoabeans.scoreboard.packet.Packet;
+import net.apartium.cocoabeans.packet.Packet;
 
 public record RemoveTeamPacket() implements Packet {
 }