+++ scoreboard/src/test/java/net/apartium/cocoabeans/scoreboard/display/packet/UpdateTeamPacket.java
@@ -1,6 +1,6 @@
 package net.apartium.cocoabeans.scoreboard.display.packet;
 
-import net.apartium.cocoabeans.scoreboard.packet.Packet;
+import net.apartium.cocoabeans.packet.Packet;
 import net.apartium.cocoabeans.scoreboard.team.ChatFormatting;
 import net.apartium.cocoabeans.scoreboard.team.CollisionRule;
 import net.apartium.cocoabeans.scoreboard.team.NameTagVisibilityRule;