+++ scoreboard/src/test/java/net/apartium/cocoabeans/scoreboard/packet/ScorePacket.java
@@ -1,5 +1,6 @@
 package net.apartium.cocoabeans.scoreboard.packet;
 
+import net.apartium.cocoabeans.packet.Packet;
 import net.apartium.cocoabeans.scoreboard.ScoreboardAction;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.Style;